Skip and Delete During Rank Pause
The indices are back to objects again, no more simple ints. I did this so I could track whether the character got marked as skip or not without having to mess around with the indices. Resume now takes a count so the size can be updated appropriately. Resume handles taking out all of the skipped characters from the indices. Added a sortIndices method that converts the indices array to a 1...n sorted list of the same length. Merged the logic for removeIndex and removeCurrent and handled delete (permanent). Added the ability to send an index in to addIndex so the indices array would stay accurate. You can now mark an index as skip, which gets handled in Resume as mentioned.